+++ ceammc/ext/src/midi/midi_octave.h
@@ -26,7 +26,7 @@ public:
     MidiOctave(const PdArgs& args);
     void onFloat(t_float note) override;
     void onList(const AtomList& lst) override;
-    void onInlet(size_t n, const AtomListView& lv) override;
+    void onInlet(size_t, const AtomListView& lv) override;
 
 private:
     t_float octave(t_float note) const;